+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [2.2.0-next.2](https://github.com/warp-ds/icons/compare/v2.2.0-next.1...v2.2.0-next.2) (2024-10-17)
+
+
+### Features
+
+* add trend-down/flat/up icons ([#131](https://github.com/warp-ds/icons/issues/131)) ([4350df9](https://github.com/warp-ds/icons/commit/4350df92cbeaaa60125cdc266edd32f4c7bf55c4))
+
+# [2.2.0-next.1](https://github.com/warp-ds/icons/compare/v2.1.0...v2.2.0-next.1) (2024-10-16)
+
+
+### Features
+
+* add Sparkles and Football icon ([#130](https://github.com/warp-ds/icons/issues/130)) ([52a5f08](https://github.com/warp-ds/icons/commit/52a5f08c89af5eec13817fa67f04cd5173d2403b))
+
 # [2.1.0](https://github.com/warp-ds/icons/compare/v2.0.2...v2.1.0) (2024-08-15)